+++ src/js/animx.js
@@ -36,7 +36,7 @@ import { refreshScrollMetrics, clearScrollTicker } from './scroll/scroll-ticker.
 import { svg, svgDraw, svgUndraw, svgProgress, svgPathFollow, bindSvgAnimX, destroySvg } from './svg/svg-api.js';
 import { cms, refreshCMS, observeCMS, disconnectCMS, getCMSRecipes, applyRecipe, bindCMSApi } from './cms/cms-api.js';
 
-const VERSION = '2.1.0';
+const VERSION = '2.2.0';
 
 // Pre-register all presets
 [...Object.values(cssPresets), ...componentPresets, ...expandedPresets].forEach(preset => {